Replace POST /commands
