+++ comms/Dockerfile
@@ -8,4 +8,4 @@ COPY . .
 
 EXPOSE 5000
 
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "5000"]
+ENTRYPOINT ["python3", "main.py"]
