--- comms/indexer_mock/Dockerfile
@@ -1,9 +1,0 @@
-FROM python:3.10-alpine
-
-COPY requirements.txt .
-
-RUN pip install --no-cache-dir --upgrade -r requirements.txt
-
-COPY . .
-
-ENTRYPOINT ["python3", "./main.py"]
