--- comms/indexer_mock/requirements.txt
@@ -1,1 +1,0 @@
-requests
